# Vectorize advanced pairwise CD