Speedup on macOS JDK16 i7:
pairwise CD and distance storing: 17%
pairwise CD only: 47%

Note that this introduces vector masks on the default path. Hence, currently, there is object creation on the default path in every loop iteration (and the associated GC overhead). Once masking is better supported on platforms with native masks (AVX512 SVE), I expect these improvements to become much more drastic.

Before:

[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark advanced pairwise collision detection bench took 5193 +/- 136 ms per macro iteration
                         or 192567 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark advanced pairwise collision detection (check only) bench took 3911 +/- 300 ms per macro iteration
                         or 255667 calls per second on average
After:

[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark advanced pairwise collision detection bench took 4428 +/- 14 ms per macro iteration
                         or 225836 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark advanced pairwise collision detection (check only) bench took 2645 +/- 17 ms per macro iteration
                         or 378119 calls per second on average